pyproject.toml
[project]
name = "legal-multimodal-rag"
version = "0.1.0"
description = "JusticeAI - Multimodal RAG for UK Legal Caseworkers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.7.4",
    "lancedb>=0.8.0",
    "rank-bm25>=0.2.2",
    "xgboost>=2.0.0",
    "scikit-learn>=1.4.0",
    "streamlit>=1.35.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "diskcache>=5.6.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "Pillow>=10.3.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "tenacity>=8.3.0",
    "structlog>=24.1.0",
    "pyarrow>=15.0.0",
    "lxml>=5.2.0",
]
